backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
redis>=4.0.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.6
aiohttp>=3.8.0
requests>=2.31.0
